Buffers, error display
Buffers can now be registered, acquired and released. They are drawn to the right
of the time line.

Exceptions in threads are now printed with thread, and the timeline is
stopped. Global exceptions (e.g. no progress) are printed at the top.

Changed the sample code a bit to demonstrate buffers.